parser.py ready for api usage
