+++ data_provider/README.txt
@@ -1,3 +1,46 @@
+Переделал parser. Исправил ошибки, убрал ненужное, временно объединил в один скрипт raw_data_provider и parser,
+запустив raw_data_provider в отдельном потоке.
+
+Реализовал примитивное api в модуле api.py:
+	http://127.0.0.1:5000/api (GET):
+		 Возвращает UID (0) для получения писем, пришедших после этого UID.
+	http://127.0.0.1:5000/api (POST):
+		 Принимает на вход json со списком писем (каждое из них - словарь).
+
+Чтобы попробовать, потребуется:
+1. Модули pip (в requirements.txt):
+	pyzmq
+	msgpack
+	Flask
+	requests
+
+2. Модуль "imap_credentials.py":
+	Этот файл добавлен в .gitignore, чтобы случайно не залить в репу свои логин и пароль от почты.
+	Внутри должны быть 2 строки:
+		imap_username = ''
+		imap_password = ''
+	Внутри одинарных ковычек логин и пароль соответственно.
+
+Далее запускаете api.py, затем parser.py. При этом:
+1. В терминале с api.py можно будет наблюдать список тем и вложений
+пришедших писем.
+2. В терминале с parser.py можно будет наблюдать сообщения
+про связь с сервером api.py и результатами опроса imap сервера.
+
+Таким образом parser.py полностью готов для использования с
+api django, надо будет только поменять URL в начале файла
+"new_parser.py" на URL django, когда там будет реализовано api.
+
+
+
+
+
+
+
+
+Содержание предыдущего README.txt ниже для истории:
+
+
 На прошлой встрече создали черновой дизайн приложения. Передо мной стояло 2 задачи:
 	№15 "Вынести отправитель и парсер в отдельные модули"
 	№17 "Очередь сделать для взаимодействия сервера и парсера"
